+++ src/__TESTS__/footer.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Footer from '../components/comp/Footer';
+import Footer from '../components/component/Footer';
 
 it('renders Footer correctly', () => {
   const tree = renderer.create(<Footer />);
