+++ src/css/bookAppointmen.module.css
@@ -35,7 +35,8 @@
 .aside span:last-child {
   width: 100%;
   display: flex;
-  justify-content: center;
+  flex-direction: column;
+  align-items: center;
 }
 
 .aside input {
