+++ src/css/index.css
@@ -45,8 +45,6 @@ code {
   position: fixed;
   top: 0;
   left: 0;
-
-  /* left:-100% ; */
   display: flex;
   flex-direction: column;
   padding: 20px 0 20px 20px;
